Modal: Renames insert/edit value modal
Upsert is a term from SQL that is either an insert or update, so
I think it is an appropriate term to use for a modal that can be
used to either add or edit a value.

Also removes the 'saw'-prefix from the file and class name, since
conflicts there are unlikely. Keeps it in the defined HTML-tag
though since conflicts there are a bit more likely.