Adds basic value info modal
Currently unused. Meant for displaying additional information about
a value so the value cards aren't so cluttered.

The information to be displayed in the info modal is: name,
description, notes and a list of habits connected to this value.

We could also consider moving the edit/delete value buttons into
the info modal, as these operations should not be used frequently
enough to warrant taking up space in the value cards.